+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/validate/__init__.py
@@ -6,7 +6,6 @@ import click
 from ..console import CONTEXT_SETTINGS
 from .agent_reqs import agent_reqs
 from .agent_signature import legacy_signature
-from .all_validations import all
 from .codeowners import codeowners
 from .config import config
 from .dashboards import dashboards
@@ -14,9 +13,7 @@ from .dep import dep
 from .eula import eula
 from .imports import imports
 from .integration_style import integration_style
-from .jmx_metrics import jmx_metrics
 from .license_headers import license_headers
-from .licenses import licenses
 from .models import models
 from .package import package
 from .readmes import readmes
@@ -26,7 +23,6 @@ from .typos import typos
 
 ALL_COMMANDS = (
     agent_reqs,
-    all,
     codeowners,
     config,
     dashboards,
@@ -34,10 +30,8 @@ ALL_COMMANDS = (
     eula,
     imports,
     integration_style,
-    jmx_metrics,
     legacy_signature,
     license_headers,
-    licenses,
     models,
     package,
     readmes,