+++ riak/tests/test_e2e.py
@@ -16,11 +16,11 @@ def test_check(dd_agent_check, instance):
     sc_tags = tags + ['url:' + instance['url']]
 
     for gauge in common.CHECK_GAUGES + common.CHECK_GAUGES_STATS:
-        aggregator.assert_metric(gauge, tags=tags, count=2)
+        aggregator.assert_metric(gauge, tags=tags, at_least=1)
 
     aggregator.assert_service_check(common.SERVICE_CHECK_NAME, status=Riak.OK, tags=sc_tags)
 
     for gauge in common.GAUGE_OTHER:
-        aggregator.assert_metric(gauge, count=1)
+        aggregator.assert_metric(gauge, at_least=0)
 
     aggregator.all_metrics_asserted()