+++ amazon_msk/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 <!-- towncrier release notes start -->
 
+## 7.7.0 / 2026-04-15
+
+***Added***:
+
+* Update dependencies ([#22996](https://github.com/DataDog/integrations-core/pull/22996))
+
+***Fixed***:
+
+* Improve descriptions ([#23047](https://github.com/DataDog/integrations-core/pull/23047))
+
+## 7.6.0 / 2026-04-01 / Agent 7.78.0
+
+***Added***:
+
+* Add support for security validation in models ([#23109](https://github.com/DataDog/integrations-core/pull/23109))
+
 ## 7.5.0 / 2026-03-18
 
 ***Added***: