+++ kubevirt_api/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 <!-- towncrier release notes start -->
 
+## 2.5.1 / 2026-04-15
+
+***Fixed***:
+
+* Improve descriptions ([#23047](https://github.com/DataDog/integrations-core/pull/23047))
+
+## 2.5.0 / 2026-04-01 / Agent 7.78.0
+
+***Added***:
+
+* Add support for security validation in models ([#23109](https://github.com/DataDog/integrations-core/pull/23109))
+
 ## 2.4.0 / 2026-03-18
 
 ***Added***: